Use a socket timeout of 1 second when using Python's builtin SSL
ntske-client.py keeps the socket open to see if a server sends
anything past the EOM.  time.cloudflare.com doesn't close the socket
after sending the EOM record which means that the client will
hang forever without a timeout.  This is only visible with
the builtin SSL library because the PyOpenSSL implementation
can't do TLSv1.3 and cloudflare requires TLSv1.3.